Remove `allow_wildcard` from JSON API includes.
Wildcards are not part of the specification and they
were explicitly stated to be an addition to the
specification by the makers of `jsonapi-renderer`. [1]

As such, supporting them seems to be not that much
of a prudent idea, given that they are not part of the
unreleased v1.1 specification, either.

[1] https://github.com/jsonapi-rb/jsonapi-renderer/blob/master/lib/jsonapi/include_directive.rb